live_data_fetching_api/requirements.txt
pytz==2022.7.1
requests==2.28.2
psycopg2-binary==2.9.5
